Reorg secondary feedback files.
